START: rewriting iterative runlength encoder - switch back to encoding all info - need to track missing
Because continuous encoding doens't get a linear sweep of all offsets in encoding range, so it can see which are not actively in the incomplete list, we need to track both positive and negative, so that missing offsets can be interpreted as missing, instead of assuming they're incomplete. Contract requires that any missing data is encoded as succeeded, so that we don't load it in the incomplete map an track it as an explicitly incomplete record. However, now that we truncate incomplete data, based on what's missing in the polled batches, this would still actually work. However, the implementation where we record both is more explicit and easier to understand and more cohesive.